+++ lib/bloom_net_center_client/models/txn.rb
@@ -9,13 +9,13 @@ module BloomNetCenterClient
     attribute :sender_street, String
     attribute :sender_city, String
     attribute :sender_province, String
-    attribute :sender_contact_number, String
+    attribute :sender_mobile, String
     attribute :recipient_first_name, String
     attribute :recipient_last_name, String
     attribute :recipient_street, String
     attribute :recipient_city, String
     attribute :recipient_province, String
-    attribute :recipient_contact_number, String
+    attribute :recipient_mobile, String
     attribute :amount, BigDecimal
     attribute :account, String
     attribute :currency, String