+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unrelased]
+### Changed
+- Rename `contact_number` to `mobile` (follow center's new API)
+
 ## [0.3.1] - 2017-03-31
 ### Fixed
 - FindTxnResponse: `txn` is nil (no longer raises exception) when the response is invalid